Moving the batch commit control to DailyProcessToCreateHistory
Reason: There are pre-looping and post-looping data initialization and
clearing in stored DailyProcessToCreateHistory, if the batch commit is
in 6-wwi-data-simulation.sql, there will be duplicated pre-loop and
wrong data reinitialization.

Solution: Move the batch commit into DailyProcessToCreateHistory's WHILE
loop. And

Known issue: the day after 20151220 will trigger
RecordColdRoomTemperatures, which is slow and it is the next work item
to tune.